requirement.txt
apturl==0.5.2
beautifulsoup4==4.4.1
blinker==1.3
Brlapi==0.6.4
chardet==3.0.4
checkbox-support==0.22
command-not-found==0.3
cryptography==1.2.3
cycler==0.9.0
decorator==4.0.6
defer==1.0.6
dnspython==1.15.0
dnspython3==1.15.0
feedparser==5.1.3
filemagic==1.6
guacamole==0.9.2
html5lib==0.999
httplib2==0.9.1
idna==2.6
ipwhois==0.10.3
Jinja2==2.8
language-selector==0.1
louis==2.6.4
lxml==3.5.0
machinae==1.4.0
Mako==1.0.3
MarkupSafe==0.23
matplotlib==1.5.1
numexpr==2.4.3
numpy==1.11.0
oauthlib==1.0.3
onboard==1.2.0
padme==1.1.1
pandas==0.17.1
pexpect==4.0.1
Pillow==3.1.2
ptyprocess==0.5
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.2.0
python-apt==1.1.0b1+ubuntu0.16.4.9
python-dateutil==2.7.2
python-debian==0.1.27
python-systemd==231
pytz==2018.4
pyxdg==0.25
PyYAML==3.12
relatime==1.1.0
reportlab==3.3.0
requests==2.9.1
scipy==0.17.0
sessioninstaller==0.0.0
six==1.11.0
ssh-import-id==5.5
stopit==1.1.2
system-service==0.3
tables==3.2.2
tzlocal==1.5.1
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.13.1
usb-creator==0.3.0
virtualenv==15.0.1
xdiagnose==3.8.4.1
xkit==0.0.0
yara-python==4.0.1
